move CodeLocation into MethodTrace as only used from here
