+++ lib/delfos/method_trace/code_location/filename_helpers.rb
