+++ lib/delfos/method_trace/code_location/method.rb
